+++ src/components/DatePicker/DayLabel/DayLabel.types.ts
@@ -1,0 +1,3 @@
+export type DayLabelProps = {
+  show: boolean;
+};
