+++ src/components/DatePicker/Toolbar/Toolbar.types.ts
@@ -1,0 +1,11 @@
+import { Dayjs } from 'dayjs';
+
+export type ToolbarProps = {
+  onPreviousMonth: () => void;
+  onNextMonth: () => void;
+  showYearPicker: boolean;
+  onShowYearPicker: () => void;
+  showMonthPicker: boolean;
+  onShowMonthPicker: () => void;
+  currentDay: Dayjs;
+};
